ネストされたクラスを外に出す：RendererCommonValues
