+++ Dev/Editor/EffekseerCore/Binary/RendererCommonValues.cs
@@ -1,10 +1,7 @@
 ﻿using System;
 using System.Collections.Generic;
-using System.Linq;
-using Effekseer.Data;
-using Effekseer.Data.Value;
+using Effekseer.Binary.RenderData;
 using Effekseer.Utl;
-
 
 namespace Effekseer.Binary
 {
@@ -303,506 +300,16 @@ namespace Effekseer.Binary
 			return data.ToArray().ToArray();
 		}
 
-		static float[] GetUniformsVertexes(Tuple35<MaterialFileParameter.ValueStatus, MaterialInformation.UniformInformation> tuple35)
-		{
-			float[] floats1 = new float[4];
-
-			if (tuple35.Item1 == null)
-			{
-				floats1 = tuple35.Item2.DefaultValues.ToArray();
-			}
-			else if (tuple35.Item1.Value is Float float1)
-			{
-				floats1[0] = float1.Value;
-			}
-			else if (tuple35.Item1.Value is Vector2D float2)
-			{
-				floats1[0] = float2.X.Value;
-				floats1[1] = float2.Y.Value;
-			}
-			else if (tuple35.Item1.Value is Data.Value.Vector3D float3)
-			{
-				floats1[0] = float3.X.Value;
-				floats1[1] = float3.Y.Value;
-				floats1[2] = float3.Z.Value;
-			}
-			else if (tuple35.Item1.Value is Vector4D float4)
-			{
-				floats1[0] = float4.X.Value;
-				floats1[1] = float4.Y.Value;
-				floats1[2] = float4.Z.Value;
-				floats1[3] = float4.W.Value;
-			}
-
-			return floats1;
-		}
 
 		public static byte[] GetUVBytes(TextureInformation _TexInfo,
-						Data.Value.Enum<Data.RendererCommonValues.UVType> _UVType,
-						Data.RendererCommonValues.UVFixedParamater _Fixed,
-						Data.RendererCommonValues.UVAnimationParamater _Animation,
-						Data.RendererCommonValues.UVScrollParamater _Scroll,
-						Data.RendererCommonValues.UVFCurveParamater _FCurve)
+			Data.Value.Enum<Data.RendererCommonValues.UVType> _UVType,
+			Data.RendererCommonValues.UVFixedParamater _Fixed,
+			Data.RendererCommonValues.UVAnimationParamater _Animation,
+			Data.RendererCommonValues.UVScrollParamater _Scroll,
+			Data.RendererCommonValues.UVFCurveParamater _FCurve)
 		{
 			var serializer = new AdvancedUvSerializer(_UVType, _Fixed, _Animation, _Scroll, _FCurve);
 			return serializer.SerializeUv(_TexInfo);
 		}
-
-		private sealed class BasicUvSerializer : UvSerializer
-		{
-			private readonly Data.RendererCommonValues.UVAnimationSupportedFrameBlendParameter _animationParameter;
-
-			public BasicUvSerializer(Data.RendererCommonValues value)
-				: base(value.UV, value.UVFixed, value.UVAnimation.AnimationParams, value.UVScroll, value.UVFCurve)
-			{
-				_animationParameter = value.UVAnimation;
-			}
-
-			protected override void GetSize(TextureInformation texInfo, out float width, out float height)
-			{
-				// specification change(1.5)
-				width = 128.0f;
-				height = 128.0f;
-
-				if (texInfo.Width > 0 && texInfo.Height > 0)
-				{
-					width = texInfo.Width;
-					height = texInfo.Height;
-				}
-			}
-
-			public override void SerializeAnimationUv(UvAggregator aggregator)
-			{
-				base.SerializeAnimationUv(aggregator);
-				aggregator.Add(_animationParameter.FlipbookInterpolationType);
-			}
-		}
-
-		private sealed class AdvancedUvSerializer : UvSerializer
-		{
-			public AdvancedUvSerializer(
-				Enum<Data.RendererCommonValues.UVType> uvType,
-				Data.RendererCommonValues.UVFixedParamater @fixed,
-				Data.RendererCommonValues.UVAnimationParamater animation,
-				Data.RendererCommonValues.UVScrollParamater scroll,
-				Data.RendererCommonValues.UVFCurveParamater fCurve)
-				: base(uvType, @fixed, animation, scroll, fCurve)
-			{
-			}
-
-			protected override void GetSize(TextureInformation texInfo, out float width, out float height)
-			{
-				// specification change(1.5)
-				width = 128.0f;
-				height = 128.0f;
-
-				if (texInfo.Width > width && texInfo.Height > height)
-				{
-					width = texInfo.Width;
-					height = texInfo.Height;
-				}
-			}
-		}
-
-		private abstract class UvSerializer
-		{
-			private Enum<Data.RendererCommonValues.UVType> UvType { get; }
-			private Data.RendererCommonValues.UVFixedParamater Fixed { get; }
-			private Data.RendererCommonValues.UVAnimationParamater Animation { get; }
-			private Data.RendererCommonValues.UVScrollParamater Scroll { get; }
-			private Data.RendererCommonValues.UVFCurveParamater FCurve { get; }
-
-			protected UvSerializer(Enum<Data.RendererCommonValues.UVType> uvType,
-				Data.RendererCommonValues.UVFixedParamater @fixed,
-				Data.RendererCommonValues.UVAnimationParamater animation,
-				Data.RendererCommonValues.UVScrollParamater scroll,
-				Data.RendererCommonValues.UVFCurveParamater fCurve)
-			{
-				UvType = uvType;
-				Fixed = @fixed;
-				Animation = animation;
-				Scroll = scroll;
-				FCurve = fCurve;
-			}
-
-			public byte[] SerializeUv(TextureInformation texInfo)
-			{
-				GetSize(texInfo, out float width, out float height);
-
-				var aggregator = new UvAggregator(width, height);
-				aggregator.Add(UvType);
-
-				switch (UvType.Value)
-				{
-					case Data.RendererCommonValues.UVType.Default:
-						break;
-					case Data.RendererCommonValues.UVType.Fixed:
-						SerializeFixedUv(aggregator);
-						break;
-					case Data.RendererCommonValues.UVType.Animation:
-						SerializeAnimationUv(aggregator);
-						break;
-					case Data.RendererCommonValues.UVType.Scroll:
-						SerializeScrollUv(aggregator);
-						break;
-					case Data.RendererCommonValues.UVType.FCurve:
-						SerializeFCurveUv(aggregator);
-						break;
-				}
-
-				return aggregator.CurrentData.ToArray().ToArray();
-			}
-
-			protected abstract void GetSize(TextureInformation texInfo, out float width, out float height);
-
-			public virtual void SerializeFCurveUv(UvAggregator aggregator)
-			{
-				var width = aggregator.Width;
-				var height = aggregator.Height;
-				aggregator.Add(FCurve.Start.GetBytes(1.0f / width, 1.0f / height));
-				aggregator.Add(FCurve.Size.GetBytes(1.0f / width, 1.0f / height));
-			}
-
-			public virtual void SerializeFixedUv(UvAggregator aggregator)
-			{
-				aggregator.AddVector2D(Fixed.Start);
-				aggregator.AddVector2D(Fixed.Size);
-			}
-
-			public virtual void SerializeAnimationUv(UvAggregator aggregator)
-			{
-				aggregator.AddVector2D(Animation.Start);
-				aggregator.AddVector2D(Animation.Size);
-
-				var frameLength = Animation.FrameLength.Infinite
-					? (int.MaxValue / 100)
-					: Animation.FrameLength.Value.Value;
-				aggregator.AddInt(frameLength);
-
-				aggregator.AddInt(Animation.FrameCountX.Value);
-				aggregator.AddInt(Animation.FrameCountY.Value);
-				aggregator.Add(Animation.LoopType);
-
-				aggregator.AddInt(Animation.StartSheet.Max);
-				aggregator.AddInt(Animation.StartSheet.Min);
-			}
-
-
-			public virtual void SerializeScrollUv(UvAggregator aggregator)
-			{
-				aggregator.AddRandomVector2D(Scroll.Start);
-				aggregator.AddRandomVector2D(Scroll.Size);
-				aggregator.AddRandomVector2D(Scroll.Speed);
-			}
-		}
-
-		private sealed class UvAggregator
-		{
-			public float Width { get; }
-			public float Height { get; }
-			private readonly List<byte[]> _data = new List<byte[]>();
-
-			public IEnumerable<byte[]> CurrentData => _data;
-
-			public UvAggregator(float width, float height)
-			{
-				Width = width;
-				Height = height;
-			}
-
-			public void AddInt(int value)
-			{
-				_data.Add(value.GetBytes());
-			}
-
-			public void AddFloat(float value)
-			{
-				_data.Add(value.GetBytes());
-			}
-
-			public void Add(byte[] value)
-			{
-				_data.Add(value);
-			}
-
-			public void AddVector2D(Vector2D vector)
-			{
-				AddFloat(vector.X / Width);
-				AddFloat(vector.Y / Height);
-			}
-
-			public void AddRandomVector2D(Vector2DWithRandom vector)
-			{
-				AddFloat(vector.X.Max / Width);
-				AddFloat(vector.Y.Max / Height);
-				AddFloat(vector.X.Min / Width);
-				AddFloat(vector.Y.Min / Height);
-			}
-		}
-
-		private sealed class MaterialSerializer
-		{
-			private void AddTextureInfo(ExporterVersion version,
-				Dictionary<string, int> texAndInd,
-				TextureValuesAggregator aggregator)
-			{
-				if (version < ExporterVersion.Ver16Alpha1)
-				{
-					return;
-				}
-
-				aggregator.AddAlphaTexture(texAndInd);
-				aggregator.AddUvDistortionTexture(texAndInd);
-
-				if (aggregator.IsBlendTextureEnabled)
-				{
-					aggregator.AddBlendTexture(texAndInd);
-					aggregator.AddBlendAlphaTexture(texAndInd);
-					aggregator.AddBlendUvDistortionTexture(texAndInd);
-				}
-				else
-				{
-					aggregator.AddInt(-1);
-					aggregator.AddInt(-1);
-					aggregator.AddInt(-1);
-				}
-			}
-
-			private void AddAsDefaultMaterial(ExporterVersion version,
-				Dictionary<string, int> texture_and_index,
-				Data.RendererCommonValues value,
-				TextureValuesAggregator aggregator)
-			{
-				aggregator.AddTexIdAndStoreSize(value.ColorTexture, 1, texture_and_index);
-				aggregator.AddInt(-1);
-				AddTextureInfo(version, texture_and_index, aggregator);
-			}
-
-			private void AddAsBackDistortion(ExporterVersion version,
-				Dictionary<string, int> distortionTexture_and_index,
-				Data.RendererCommonValues value,
-				TextureValuesAggregator aggregator)
-			{
-				aggregator.AddTexIdAndStoreSize(value.ColorTexture, 1, distortionTexture_and_index);
-				aggregator.AddInt(-1);
-				AddTextureInfo(version, distortionTexture_and_index, aggregator);
-			}
-
-			private void AddAsLighting(ExporterVersion version,
-				Dictionary<string, int> texture_and_index,
-				Dictionary<string, int> normalTexture_and_index,
-				Data.RendererCommonValues value,
-				TextureValuesAggregator aggregator)
-			{
-				aggregator.AddTexIdAndStoreSize(value.ColorTexture, 1, texture_and_index);
-				aggregator.AddTexIdAndStoreSize(value.NormalTexture, 2, normalTexture_and_index);
-				AddTextureInfo(version, texture_and_index, aggregator);
-			}
-
-			private void AddAsFile(TextureValuesAggregator aggregator,
-				Data.RendererCommonValues value,
-				Dictionary<string, int> texture_and_index,
-				Dictionary<string, int> normalTexture_and_index,
-				Dictionary<string, int> material_and_index)
-			{
-				var materialInfo = Core.ResourceCache.LoadMaterialInformation(value.MaterialFile.Path.AbsolutePath) ??
-					new MaterialInformation();
-				var textures = value.MaterialFile.GetTextures(materialInfo).Where(_ => _.Item1 != null).ToArray();
-				var uniforms = value.MaterialFile.GetUniforms(materialInfo);
-
-				// maximum slot limitation
-				if (textures.Length > Constant.UserTextureSlotCount)
-				{
-					textures = textures.Take(Constant.UserTextureSlotCount).ToArray();
-				}
-
-				aggregator.AddInt(material_and_index.ContainsKey(value.MaterialFile.Path.RelativePath)
-					? material_and_index[value.MaterialFile.Path.RelativePath]
-					: (-1));
-
-				aggregator.AddInt(textures.Length);
-
-				foreach (var texture in textures)
-				{
-					var texture_ = texture.Item1.Value as Data.Value.PathForImage;
-					if (texture.Item2.Type == TextureType.Value)
-					{
-						aggregator.AddInt(1);
-						aggregator.AddTexIdAndStoreSize(texture_, texture.Item2.Priority, normalTexture_and_index);
-					}
-					else
-					{
-						aggregator.AddInt(0);
-						aggregator.AddTexIdAndStoreSize(texture_, texture.Item2.Priority, texture_and_index);
-					}
-				}
-
-				aggregator.AddInt(uniforms.Count);
-
-				foreach (var uniform in uniforms)
-				{
-					float[] floats = GetUniformsVertexes(uniform);
-
-					aggregator.AddFloat(floats[0]);
-					aggregator.AddFloat(floats[1]);
-					aggregator.AddFloat(floats[2]);
-					aggregator.AddFloat(floats[3]);
-				}
-			}
-
-			public void AddMaterialData(ExporterVersion version,
-				Data.RendererCommonValues value,
-				TextureValuesAggregator aggregator,
-				Dictionary<string, int> texture_and_index,
-				Dictionary<string, int> distortionTexture_and_index,
-				Dictionary<string, int> normalTexture_and_index,
-				Dictionary<string, int> material_and_index)
-			{
-				switch (value.Material.Value)
-				{
-					case Data.RendererCommonValues.MaterialType.Default:
-						AddAsDefaultMaterial(version, texture_and_index, value, aggregator);
-						break;
-					case Data.RendererCommonValues.MaterialType.BackDistortion:
-						AddAsBackDistortion(version, distortionTexture_and_index, value, aggregator);
-						break;
-					case Data.RendererCommonValues.MaterialType.Lighting:
-						AddAsLighting(version, texture_and_index, normalTexture_and_index, value, aggregator);
-						break;
-					default:
-						AddAsFile(aggregator, value, texture_and_index, normalTexture_and_index, material_and_index);
-						break;
-				}
-			}
-		}
-
-		private sealed class TextureValuesAggregator
-		{
-			private readonly Data.RendererCommonValues _value;
-			private readonly AdvancedRenderCommonValues _advanceValue;
-			private readonly AdvancedRenderCommonValues2 _advanceValue2;
-			private TextureInformation _texInfo;
-			private TextureInformation _alpha;
-			private TextureInformation _uvDistortion;
-			private TextureInformation _blend;
-			private TextureInformation _blendAlpha;
-			private TextureInformation _blendUvDistortion;
-			private readonly List<byte[]> _data = new List<byte[]>();
-
-			public TextureValuesAggregator(
-				Data.RendererCommonValues value,
-				AdvancedRenderCommonValues advanceValue,
-				AdvancedRenderCommonValues2 advanceValue2,
-				TextureInformation texInfo,
-				TextureInformation alpha,
-				TextureInformation uvDistortion,
-				TextureInformation blend,
-				TextureInformation blendAlpha,
-				TextureInformation blendUvDistortion)
-			{
-				_value = value;
-				_advanceValue = advanceValue;
-				_advanceValue2 = advanceValue2;
-				_texInfo = texInfo;
-				_alpha = alpha;
-				_uvDistortion = uvDistortion;
-				_blend = blend;
-				_blendAlpha = blendAlpha;
-				_blendUvDistortion = blendUvDistortion;
-			}
-
-			public IEnumerable<byte[]> CurrentData => _data;
-			public bool IsBlendTextureEnabled => _advanceValue2.EnableBlendTexture;
-
-			public void AddInt(int value)
-			{
-				_data.Add(value.GetBytes());
-			}
-
-			public void AddFloat(float value)
-			{
-				_data.Add(value.GetBytes());
-			}
-
-			public void AddTexIdAndStoreSize(PathForImage image, int number, Dictionary<string, int> texAndInd)
-			{
-				var tempTexInfo = new TextureInformation();
-
-				if (!texAndInd.ContainsKey(image.RelativePath) || !tempTexInfo.Load(image.AbsolutePath))
-				{
-					AddInt(-1);
-					return;
-				}
-
-				if (_value.UVTextureReferenceTarget.Value != Data.UVTextureReferenceTargetType.None
-					&& number == (int)_value.UVTextureReferenceTarget.Value)
-				{
-					_texInfo.Load(image.AbsolutePath);
-				}
-
-				AddInt(texAndInd[image.RelativePath]);
-			}
-
-			private int GetTexIdAndInfo(PathForImage image, Dictionary<string, int> texAndInd, ref TextureInformation texInfoRef)
-			{
-				var tempTexInfo = new TextureInformation();
-
-				if (!texAndInd.ContainsKey(image.RelativePath) || !tempTexInfo.Load(image.AbsolutePath))
-				{
-					return -1;
-				}
-
-				texInfoRef.Load(image.AbsolutePath);
-				return texAndInd[image.RelativePath];
-			}
-
-			private void AddTexIdAndInfo(bool isEnabled, PathForImage path, Dictionary<string, int> texAndInd, ref TextureInformation textureInfo)
-			{
-				var value = isEnabled
-					? GetTexIdAndInfo(path, texAndInd, ref textureInfo)
-					: -1;
-				AddInt(value);
-			}
-
-			public void AddAlphaTexture(Dictionary<string, int> texAndInd)
-			{
-				AddTexIdAndInfo(_advanceValue.EnableAlphaTexture,
-					_advanceValue.AlphaTextureParam.Texture,
-					texAndInd,
-					ref _alpha);
-			}
-
-			public void AddUvDistortionTexture(Dictionary<string, int> texAndInd)
-			{
-				AddTexIdAndInfo(_advanceValue.EnableUVDistortionTexture,
-					_advanceValue.UVDistortionTextureParam.Texture,
-					texAndInd,
-					ref _uvDistortion);
-			}
-
-			public void AddBlendAlphaTexture(Dictionary<string, int> texAndInd)
-			{
-				AddTexIdAndInfo(_advanceValue2.BlendTextureParams.EnableBlendAlphaTexture,
-					_advanceValue2.BlendTextureParams.BlendAlphaTextureParam.Texture,
-					texAndInd,
-					ref _blendAlpha);
-			}
-
-			public void AddBlendUvDistortionTexture(Dictionary<string, int> texAndInd)
-			{
-				AddTexIdAndInfo(_advanceValue2.BlendTextureParams.EnableBlendUVDistortionTexture,
-					_advanceValue2.BlendTextureParams.BlendUVDistortionTextureParam.Texture,
-					texAndInd,
-					ref _blendUvDistortion);
-			}
-
-			public void AddBlendTexture(Dictionary<string, int> texAndInd)
-			{
-				GetTexIdAndInfo(_advanceValue2.BlendTextureParams.BlendTextureParam.Texture,
-					texAndInd,
-					ref _blend);
-			}
-		}
 	}
-
 }
